changed type of size

+++ include/avMemory.h
@@ -8,7 +8,7 @@ extern "C" {
 #include "avTypes.h"
 
 
-void* avAllocate_(size_t size, const char* message, uint line, const char* func, const char* file);
+void* avAllocate_(uint64 size, const char* message, uint line, const char* func, const char* file);
 
 #ifdef AV_ALLOCATE_ZERO_OUT
 #define avAllocate(size, message) avCallocate_(1, size, message, __LINE__, __func__, __FILE__);
@@ -16,10 +16,10 @@ void* avAllocate_(size_t size, const char* message, uint line, const char* func,
 #define avAllocate(size, message) avAllocate_(size, message, __LINE__, __func__, __FILE__)
 #endif
 
-void* avCallocate_(size_t count, size_t size, const char* message, uint line, const char* func, const char* file);
+void* avCallocate_(uint64 count, uint64 size, const char* message, uint line, const char* func, const char* file);
 #define avCallocate(count, size, message) avCallocate_(count, size, message, __LINE__, __func__, __FILE__)
 
-void* avReallocate_(void* data, size_t size, const char* message, uint line, const char* func, const char* file);
+void* avReallocate_(void* data, uint64 size, const char* message, uint line, const char* func, const char* file);
 #define avReallocate(size, message) avReallocate_(size, message, __LINE__, __func__, __FILE__)
 
 void avFree_(void* data, uint line, const char* func, const char* file);
